CNN version to detect empty worlds
This version was the first I used to incorporate the feature of identifying non-claimed worlds. I collected images myself by taking screenshots of claimed and non-claimed worlds. I then trained the cnn model and exported it. It would then be used to screen capture live and save the name of the world if it finds that it is not claimed. Only issue is we get too many false positives with this model and collecting more data is very difficult and is quite a manual process. I switched to OCR because of these issues.